Try to understand mapping between frequencies-WIP
We know, given is,it,iu and nwpr, what we're going to access in Gamma.
But given a certain datum in Gamma, what are the values of is,it and iu
that need it? and with which value of nwpr?

This might lead us to better data reuse.

Note: inverse mapping

function mixedFrequenciesConvertedInverse(ns, t1, u1, t2, u2,
flavTransf12, flavTransf34, N)

does not yet work, except for a few cases in the test suite.
We need to find single failing test cases and fix the function from there.