mixedFrequenciesConvertedInverse working version
This inverted makes a lot of trials and error
When there is only one possible candidate in the domain,
then it is found without errors.
But when the cutoff logic starts working,
then the current version of the inverse function
produces a lot of candidates either out of bound
or for which the reconstruction of the original would fail.

Average success rate for current reconstruction algorithm
is less than 10% for N=40.

Discovered, though,
that the slices
- is in 39:40
- it in 39:40
- iu in 39:40
are accessed 50% of the times compared to the full 40^3
matsubara frequency cube,
while representing only 14% of the whole data set
(in Gamma)